knowledge_slanted

The knowledge-slanted RF approach to identify genes potentially implicated with CAVS in patients with congenital bicuspid aortic valve (BAV) and tricuspid aortic valve (TAV) in comparison with patients having normal valves, using a knowledge-slanted random forest (RF).

knowledge_slanted(data = NULL, ntree, mtry, weights, ...)

Arguments

data

Training data of class data.frame, matrix

ntree

Number of trees.

mtry

Number of variables to possibly split at in each node.

weights

Numeric vector with weights between 0 and 1

...

further arguments to be passed to or from other methods especially

test = data.frame, to predict the values based on the previous data

do.metrics = TRUE, to list different metrics that were added to your model (error S E auc F1)

Details

The knowledge-slanted RF approach to identify genes potentially implicated with CAVS in patients

References

Biological knowledge-slanted random forest approach for the classification of calcified aortic valve stenosis 10.1186/s13040-021-00269-4

See also