ksboruta

Boruta is an all relevant feature selection wrapper algorithm, boruta_method is a wrapper for knowledge-slanted random forest implementation. It tries to pick all the important dataset features

ksboruta(data, ntree, mtry, weights, trace = FALSE, ...)

Arguments

data

Training data of class data.frame or matrix

ntree

integer(1) number of tree.

mtry

integer(1) number of Variables randomly chosen at each split

weights

numeric vector with weights between 0 and 1.

trace

verbose. 0 means no tracing, 1 means decision about each attribute, 2 means the same as 1, plus reporting step by step,

...

further arguments passed to ranger().

Value

ks object, see knowledge_slanted() for details.

Details

Runs a feature selection function

References

Biological knowledge-slanted random forest approach for the classification of calcified aortic valve stenosis 10.1186/s13040-021-00269-4

See also